kgplus: colocate TransmissionKeys with its serialization
Move WriteTo/ReadFrom/BinarySize/MarshalBinary/UnmarshalBinary from
kgplus/serialization.go into kgplus/keys.go next to the TransmissionKeys
struct, then delete kgplus/serialization.go. Matches the masterkey.go
and lattigo conventions: type and methods together.
